Create query parser to capture from #annotations.
The process I am currently working with is to accept a value
from the #annotations: and then detect query__.

Once this occurs, the rest of the statement will be used as a
query in Hypothesis's search client.  Statements like "user",
"tag" or "any" followed by a double underscore (user__) will
be converted into a search query identifier.

I am aware of a secondary solution that could accept #query:
(or some such) but that appears to require changes to the
browser extension, which I have wanted to avoid. However,
I will include a mockup for this change in separate branches.